requirements.txt
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
numpy==1.14.5
pandas==0.24.0
scikit-learn==0.20.2
scipy==1.1.0
SQLAlchemy==1.2.17
tensorflow-gpu==1.12.0
